AIMS - refactor handlers
